Chore: 배포를 위해 auth/[…nextauth]/route.ts 수정
